Move api to lib
